// src/app/zdslogic-plugins/logs/core/models/accessAbuse.model.ts
export class AccessAbuse {

	ipAddress: string;
	isPublic: boolean;
	ipVersion: number;
	isWhitelisted: boolean;
	abuseConfidenceScore:number;
	countryCode:string;
	usageType:string;
	isp:string;
	domain:string;
	hostnames:string;
	totalReports: number;
	numDistinctUsers: number;
	lastReportedAt: string;
		
}